Update files to compile in Windows, Visual Studio 2019

+++ src/slide-merger-task.cpp
@@ -54,7 +54,7 @@ void SlideMergerTask::setup(const std::vector<cv::Mat> & inputFrames, uint32_t p
 }
 
 void SlideMergerTask::execute(const cv::Mat & inputFrame, cv::Mat & outputFrame, uint32_t frameIndex, std::mutex & mutex, uint32_t passIndex) {
-    uint32_t currentImage = closestImageIndex(inputFrame, _originalImages, _currentImage);
+    uint32_t currentImage = static_cast<uint32_t>(closestImageIndex(inputFrame, _originalImages, _currentImage));
     if (currentImage!=_currentImage) {
         _slideImage = &inputFrame;
     }
